feat(projects): give each project tab its own route
Tabs on /projects/[projectId] and /share/[token] were local state, so they
weren't linkable, didn't survive a reload, and ignored the back button. Each
tab is now a route segment, with the Document Explorer staying at the root.

Both routes hold their shell in a layout so tab navigation doesn't remount it:
the project layout keeps the selected revision across tab changes, which a
per-page approach would reset to latest. Since a layout can't pass props to
`children`, ProjectViewProvider carries projectDetail, readOnly, revision state
and navigation down to the tab pages.

ResultsVisualization is split into ProjectResultsShell (chrome + tab list,
taking the active tab's content as children) and ProjectTabPanel (the tab
switch). Both routes now use the shell directly, so the old component has no
callers and is removed.
