+++ frontend/app/share/[token]/analyses/page.tsx
@@ -3,5 +3,5 @@
 import { ProjectTabPanel } from '@/components/results/project-tab-panel';
 
 export default function Page() {
-  return <ProjectTabPanel tab="document-explorer" />;
+  return <ProjectTabPanel tab="analyses" />;
 }
