+++ frontend/app/share/[token]/summary/page.tsx
@@ -3,5 +3,5 @@
 import { ProjectTabPanel } from '@/components/results/project-tab-panel';
 
 export default function Page() {
-  return <ProjectTabPanel tab="document-explorer" />;
+  return <ProjectTabPanel tab="summary" />;
 }
